# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalRequest": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "minLength": 1,
            "title": "Comment",
            "type": "string"
          },
          "reviewer": {
            "minLength": 1,
            "title": "Reviewer",
            "type": "string"
          }
        },
        "required": [
          "reviewer",
          "comment"
        ],
        "title": "ApprovalRequest",
        "type": "object"
      },
      "ApprovalResponse": {
        "additionalProperties": false,
        "properties": {
          "approval": {
            "$ref": "#/components/schemas/ApprovalSummary"
          },
          "approved": {
            "title": "Approved",
            "type": "boolean"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "approved",
          "approval"
        ],
        "title": "ApprovalResponse",
        "type": "object"
      },
      "ApprovalSummary": {
        "additionalProperties": false,
        "properties": {
          "approved_at": {
            "format": "date-time",
            "title": "Approved At",
            "type": "string"
          },
          "comment": {
            "title": "Comment",
            "type": "string"
          },
          "reviewer": {
            "title": "Reviewer",
            "type": "string"
          },
          "ruleset_version": {
            "title": "Ruleset Version",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          }
        },
        "required": [
          "reviewer",
          "comment",
          "approved_at",
          "run_id",
          "snapshot_id",
          "ruleset_version"
        ],
        "title": "ApprovalSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LockRequest": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "minLength": 1,
            "title": "Job Id",
            "type": "string"
          },
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "reason"
        ],
        "title": "LockRequest",
        "type": "object"
      },
      "LockResponse": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "locked": {
            "title": "Locked",
            "type": "boolean"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Reason Codes",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "job_id",
          "locked",
          "reason_codes"
        ],
        "title": "LockResponse",
        "type": "object"
      },
      "PlanningRunCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "ruleset_version": {
            "minLength": 1,
            "title": "Ruleset Version",
            "type": "string"
          },
          "snapshot_id": {
            "minLength": 1,
            "title": "Snapshot Id",
            "type": "string"
          }
        },
        "required": [
          "snapshot_id",
          "ruleset_version"
        ],
        "title": "PlanningRunCreateRequest",
        "type": "object"
      },
      "PlanningRunCreatedResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "ruleset_version": {
            "title": "Ruleset Version",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/PlanningRunState"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "state",
          "snapshot_id",
          "ruleset_version",
          "created_at",
          "status_url"
        ],
        "title": "PlanningRunCreatedResponse",
        "type": "object"
      },
      "PlanningRunDetail": {
        "additionalProperties": false,
        "properties": {
          "ai_estimates": {
            "items": {
              "$ref": "#/components/schemas/AiEstimate"
            },
            "title": "Ai Estimates",
            "type": "array"
          },
          "approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "changes": {
            "additionalProperties": {
              "enum": [
                "SCHEDULED",
                "REJECTED",
                "PRESERVED",
                "CHANGED"
              ],
              "type": "string"
            },
            "title": "Changes",
            "type": "object"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "export_ready": {
            "title": "Export Ready",
            "type": "boolean"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobContext"
            },
            "title": "Jobs",
            "type": "array"
          },
          "kpis": {
            "$ref": "#/components/schemas/KpiSummary"
          },
          "parent_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Run Id"
          },
          "ruleset_version": {
            "title": "Ruleset Version",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "schedule_items": {
            "items": {
              "$ref": "#/components/schemas/ScheduleItem"
            },
            "title": "Schedule Items",
            "type": "array"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          },
          "snapshot_status": {
            "title": "Snapshot Status",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/PlanningRunState"
          },
          "unscheduled_jobs": {
            "items": {
              "$ref": "#/components/schemas/UnscheduledJob"
            },
            "title": "Unscheduled Jobs",
            "type": "array"
          },
          "validator": {
            "$ref": "#/components/schemas/ValidatorSummary"
          }
        },
        "required": [
          "run_id",
          "state",
          "snapshot_id",
          "snapshot_status",
          "ruleset_version",
          "created_at",
          "completed_at",
          "schedule_items",
          "unscheduled_jobs",
          "jobs",
          "validator",
          "changes",
          "export_ready",
          "kpis",
          "ai_estimates"
        ],
        "title": "PlanningRunDetail",
        "type": "object"
      },
      "PlanningRunState": {
        "enum": [
          "QUEUED",
          "RUNNING",
          "FEASIBLE",
          "OPTIMAL",
          "INFEASIBLE",
          "TIMEOUT",
          "INVALID",
          "FAILED"
        ],
        "title": "PlanningRunState",
        "type": "string"
      },
      "PlanningRunSummary": {
        "additionalProperties": false,
        "description": "Lightweight archive row for GET /planning-runs.\n\nDeliberately excludes schedule_items/jobs/ai_estimates so listing the\narchive does not stream every job of every historical run.",
        "properties": {
          "approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "kpis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KpiSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "parent_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Run Id"
          },
          "ruleset_version": {
            "title": "Ruleset Version",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "scheduled_job_count": {
            "title": "Scheduled Job Count",
            "type": "integer"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/PlanningRunState"
          },
          "total_job_count": {
            "title": "Total Job Count",
            "type": "integer"
          },
          "trigger_type": {
            "title": "Trigger Type",
            "type": "string"
          },
          "validator_passed": {
            "title": "Validator Passed",
            "type": "boolean"
          }
        },
        "required": [
          "run_id",
          "state",
          "snapshot_id",
          "ruleset_version",
          "created_at",
          "completed_at",
          "trigger_type",
          "total_job_count",
          "scheduled_job_count",
          "validator_passed"
        ],
        "title": "PlanningRunSummary",
        "type": "object"
      },
...
      "ReplanRequest": {
        "additionalProperties": false,
        "properties": {
          "affected_section_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Affected Section Ids",
            "type": "array"
          },
          "affected_window_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Affected Window Ids",
            "type": "array"
          }
        },
        "required": [
          "affected_section_ids",
          "affected_window_ids"
        ],
        "title": "ReplanRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Decision-support API for the RailNiyojan SIH26027 vertical slice.",
    "title": "RailNiyojan API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/planning-runs": {
      "get": {
        "description": "Archive of every persisted run, newest first.",
        "operationId": "list_planning_runs_planning_runs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanningRunSummary"
                  },
                  "title": "Response List Planning Runs Planning Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Planning Runs",
        "tags": [
          "planning-runs"
        ]
      },
      "post": {
        "operationId": "create_planning_run_planning_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanningRunCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningRunCreatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Planning Run",
        "tags": [
          "planning-runs"
        ]
      }
    },
    "/planning-runs/{run_id}": {
      "get": {
        "operationId": "get_planning_run_planning_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningRunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Planning Run",
        "tags": [
          "planning-runs"
        ]
      }
    },
    "/planning-runs/{run_id}/approve": {
      "post": {
        "operationId": "approve_planning_runs__run_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve",
        "tags": [
          "planning-runs"
        ]
      }
    },
    "/planning-runs/{run_id}/export": {
      "get": {
        "operationId": "export_planning_runs__run_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export",
        "tags": [
          "planning-runs"
        ]
      }
    },
    "/planning-runs/{run_id}/lock": {
      "post": {
        "operationId": "lock_schedule_item_planning_runs__run_id__lock_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LockResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lock Schedule Item",
        "tags": [
          "planning-runs"
        ]
      }
    },
    "/planning-runs/{run_id}/replan": {
      "post": {
        "operationId": "replan_planning_runs__run_id__replan_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningRunCreatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replan",
        "tags": [
          "planning-runs"
        ]
      }
    },
...
    "/railradar/trains/{train_number}/route": {
      "get": {
        "operationId": "route_railradar_trains__train_number__route_get",
        "parameters": [
          {
            "in": "path",
            "name": "train_number",
            "required": true,
            "schema": {
              "title": "Train Number",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "geojson",
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "stops",
            "required": false,
            "schema": {
              "default": false,
              "title": "Stops",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Route Railradar Trains  Train Number  Route Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Route",
        "tags": [
          "railradar"
        ]
      }
    },
...
  }
}